requirements.txt
cycler==0.10.0
matplotlib==2.1.0
numpy==1.13.3
progressbar2==3.34.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.3
six==1.11.0